requirements.txt
flask
pandas
numpy
scikit-learn
joblib
mysql-connector-python
gunicorn
